refactor: code in cart files
