+++ src/constants/data/data.ts
@@ -15,7 +15,7 @@ export const gridValues: Array<string> = ['3', '4'];
 export const cartArray: Array<Product> = [];
 
 export const enum Pages {
-    main = '#',
-    cart = '/#cart',
-    product = '/#product-details/',
+    main = '',
+    cart = '#cart',
+    product = '#product-details/',
 }
